+++ src/rpc/core_rpc_server_commands_defs.h
@@ -2714,7 +2714,7 @@ struct request_t: public rpc_access_request_base
         uint64_t total_contributed;
         uint64_t total_reserved;
         uint64_t staking_requirement;
-        uint32_t portions_for_operator;
+        uint64_t portions_for_operator;
         std::string operator_address;
         BEGIN_KV_SERIALIZE_MAP()
           KV_SERIALIZE(service_node_pubkey)
